Update to testing file structure to match maven recommended structure of using packages to separate test classes: src/test/java/fvt and src/test/jave/unit. Move HelloWorld test to unit test.
