+++ async-jaxrs-application/pom.xml
@@ -119,15 +119,13 @@
     <properties>
     	<!-- Update this to point to a liberty server install
         This is needed to find the dependencies to build the code and to run the tests -->
-		<install.directory><!-- Set this --></install.directory>
+		<install.directory></install.directory>
 		<!-- Update this to match the port number in the JaxrsSample server.xml
 		This is needed to run the tests -->
 		<liberty.test.port.property>9081</liberty.test.port.property>
 	</properties>
 
     <build>
-	    <testSourceDirectory>${basedir}/src/fvt/java</testSourceDirectory>
-	    <testOutputDirectory>${project.build.directory}/fvt</testOutputDirectory>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -162,6 +160,7 @@
                     </execution>
                 </executions>
             </plugin>
+            <!-- Enable liberty-maven plugin so that we can start and stop a server before running tests -->
             <plugin>
             	<groupId>net.wasdev.wlp.maven.plugins</groupId>
             	<artifactId>liberty-maven-plugin</artifactId>
@@ -191,6 +190,7 @@
        				<userDirectory>${basedir}/../async-jaxrs-wlpcfg</userDirectory>
             	</configuration>
             </plugin>
+            <!-- Plugin to run functional tests -->
             <plugin>
             	<groupId>org.apache.maven.plugins</groupId>
             	<artifactId>maven-failsafe-plugin</artifactId>
@@ -201,15 +201,26 @@
             			<id>integration-test</id>
             			<goals>
             				<goal>integration-test</goal>
-            				<goal>verify</goal>
             			</goals>
             			<configuration>
-            				<testSourceDirectory>${basedir}/src/fvt/java</testSourceDirectory>
-            				<testClassesDirectory>${project.build.directory}/fvt</testClassesDirectory>
+            			    <includes>
+           						<include>**/fvt/**</include>
+           					</includes>
+           					<excludes>
+           						<exclude>**/unit/**</exclude>
+           					</excludes>
+           					<systemPropertyVariables>
+		   						<liberty.test.port>${liberty.test.port.property}</liberty.test.port>
+		      				</systemPropertyVariables>
             			</configuration>
             		</execution>
             	</executions>
+      	        <configuration>
+      				<summaryFile>${project.build.directory}/test-reports/fvt/failsafe-summary.xml</summaryFile>
+      				<reportsDirectory>${project.build.directory}/test-reports/fvt</reportsDirectory>
+      			</configuration>
             </plugin>
+            <!-- Plugin to run unit tests -->
             <plugin>
             	<groupId>org.apache.maven.plugins</groupId>
             	<artifactId>maven-surefire-plugin</artifactId>
@@ -219,20 +230,14 @@
             			<phase>test</phase>
             			<id>default-test</id>
             			<configuration>
-            				<skipTests>true</skipTests>
+           					<includes>
+           						<include>**/unit/**</include>
+           					</includes>
+           					<excludes>
+           						<exclude>**/fvt/**</exclude>
+           					</excludes>
+           					<reportsDirectory>${project.build.directory}/test-reports/unit</reportsDirectory>
             			</configuration>
-            		</execution>
-            		<execution>
-            			<phase>integration-test</phase>
-            			<id>integration-test</id>
-            			<goals>
-            				<goal>test</goal>
-            			</goals>
-           				<configuration>
-		       				<systemPropertyVariables>
-		   						<liberty.test.port>${liberty.test.port.property}</liberty.test.port>
-		      				</systemPropertyVariables>
-           				</configuration>
             		</execution>
             	</executions>
             </plugin>
