--- async-jaxrs-application/src/fvt/java/net/wasdev/jaxrs/async/test/HelloWorldTest.java
@@ -1,14 +1,0 @@
-package net.wasdev.jaxrs.async.test;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class HelloWorldTest {
-
-    @Test
-    public void runATest() {
-        assertTrue(true);
-    }
-    
-}
