+++ async-jaxrs-application/src/test/java/unit/net/wasdev/jaxrs/async/test/HelloWorldTest.java
@@ -1,0 +1,14 @@
+package unit.net.wasdev.jaxrs.async.test;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class HelloWorldTest {
+
+    @Test
+    public void runATest() {
+        assertTrue(true);
+    }
+    
+}
